FIX redundant copying of *all data* in scenefile FIX light as unique_ptr FIX placeholder checkboard texture FIX rude logic to skip unsupported texture FIX keep refraction ration as original ratio unit FIX some size_t warnings

+++ src/asset/MeshLoader.h
@@ -39,7 +39,7 @@ public:
     bool loadTextures = true;
 
 private:
-    int resolveVertexIndex(
+    size_t resolveVertexIndex(
         std::map<glm::vec3*, Vertex*>& vertexMapping,
         std::vector<Vertex*>& vertices,
         std::vector<glm::vec3>& positions,
